Cardano.Wallet.Compatibility.HttpBridge

Description

Contains various implementation decision that are specific to a particular network / protocol. This allows us to easily select a particular backend (Byron, Shelley-Rust, Shelley-Haskell) and isolate the bits that vary between those backends.

Target

data HttpBridge

A type representing the http-bridge as a network target. This has an influence on binary serializer & network primitives. See also TxId