Cardano.Wallet.HttpBridge

Description

This module provides the main wallet server function for the cardano-http-bridge backend.

Documentation

serveWallet

Arguments

:: (t ~ HttpBridge n, s ~ SeqState t, k ~ SeqKey)

=> (KeyToAddress t k, KnownNetwork n)

=> (Configuration, Switchboard Text, Trace IO Text)
Logging config.

-> Maybe FilePath
Database file.

-> Listen
HTTP API Server port.

-> HttpBridgeBackend
Whether and how to launch or use the node backend.

-> Maybe (PortNumber -> PortNumber -> NetworkLayer IO Tx (Block Tx) -> IO a)
Optional operation to run under the wallet server (e.g. test suite).

-> IO ExitCode

The cardano-wallet-http-bridge main function.