Cardano.Wallet.HttpBridge.Api

Description

An API specification for the Cardano HTTP Bridge.

Documentation

type Api = GetBlockByHash :<|> (GetEpochById :<|> (GetTipBlockHeader :<|> PostSignedTx))

type GetBlockByHash = Capture "networkName" NetworkName :> ("block" :> (Capture "blockHeaderHash" (Hash Blake2b_256 (ApiT BlockHeader)) :> Get '[CBOR] (ApiT (Block Tx))))

Retrieve a block identified by the unique hash of its header.

type GetTipBlockHeader = Capture "networkName" NetworkName :> ("tip" :> Get '[ComputeHash Blake2b_256 CBOR] (WithHash Blake2b_256 (ApiT BlockHeader)))

Retrieve the header of the latest known block.

type GetEpochById = Capture "networkName" NetworkName :> ("epoch" :> (Capture "epochId" EpochIndex :> Get '[Packed CBOR] [ApiT (Block Tx)]))

Retrieve all the blocks for the epoch identified by the given integer ID.

type PostSignedTx = Capture "networkName" NetworkName :> ("txs" :> ("signed" :> (ReqBody '[JSON] (ApiT (Tx, [TxWitness])) :> Post '[NoContent] NoContent)))

api :: Proxy Api

newtype ApiT a

Constructors

ApiT

Fields

getApiT :: a

newtype EpochIndex

Represents a unique epoch.

Constructors

EpochIndex

Fields

getEpochIndex :: Word64

newtype NetworkName

Represents the name of a Cardano network.

Constructors

NetworkName

Fields

getNetworkName :: Text