Cardano.Wallet.HttpBridge.Compatibility

Description

Contains various implementation decision that are specific to a particular network / protocol. This allows us to easily select a particular backend (Byron, Shelley-Rust, Shelley-Haskell) and isolate the bits that vary between those backends.

Target

data HttpBridge (network :: Network)

A type representing the http-bridge as a network target. This has an influence on binary serializer & network primitives. See also TxId

data Network

Available network options.

Constructors

Mainnet

Testnet

block0 :: Block Tx

An initial first block to initialize a chain using the http-bridge. We do not use the blockHash and, do only use the $sel:prevBlockHash:BlockHeader to catch up with unstable epoch and therefore, the very first $sel:prevBlockHash:BlockHeader matters not.

It isn't impossible to retrieve the blockHash by computing a blake2b 256 of the CBOR-serialized full block header, but this requires us to write the full CBOR decoders (and encoders) for the all BlockHeader which is, for the http-brdige implementation, a waste of time at the moment.

byronFeePolicy :: FeePolicy

Hard-coded fee policy for Cardano on Byron

byronSlotLength :: SlotLength

Hard-coded slot duration

byronTxMaxSize :: Quantity "byte" Word16

Hard-coded max transaction size

byronBlockchainParameters :: forall n. KnownNetwork n => BlockchainParameters (HttpBridge n)