Data.Packfile

Description

Decoder for the rust-cardano packfile format.

A pack file is a collection of bytestring blobs.

The reference implementation is in packfile.rs.

Documentation

decodePackfile :: ByteString -> Either PackfileError [ByteString]

Decode a Cardano version 1 pack file. The blobs are returned as a list. Decoding is not incremental, and all data is stored in memory.

data PackfileError

Things related to the file format that can go wrong when decoding a pack file.

Constructors

MissingMagicError

WrongFileTypeError

VersionTooOldError

VersionTooNewError

BlobDecodeError String