Servant.Extra.ContentTypes

Description

Extra content types for Servant.

Documentation

data ComputeHash algorithm a

Represents a piece of binary data for which a hash value should be calculated before performing any further deserialization.

data CBOR

Represents a CBOR (Concise Binary Object Representation) object.

See RFC 7049 (http:/cbor.io) for further details.

class FromCBOR a where

The class of types that can be converted to from CBOR.

Methods

fromCBOR :: Decoder s a

newtype Hash algorithm a

Represents the result of hashing a piece of data.

Constructors

Hash (Digest algorithm)

data Packed a

Represents something that has been packed with the Cardano packfile format.

data WithHash algorithm a

Represents a piece of data with an accompanying hash value.

Constructors

WithHash

Fields

getHash :: Digest algorithm

getValue :: a